Delete Media, Remove media FileName, Upload filtering of files

+++ VimeoCopyAPI/VimeoCopyAPI/Services/MediaService.cs
@@ -2,6 +2,7 @@
 using Amazon.S3.Model;
 using Microsoft.EntityFrameworkCore;
 using System.Net.Sockets;
+using System.Security.Claims;
 using VimeoCopyApi.Data;
 using VimeoCopyApi.Models;
 using VimeoCopyAPI.Models.DTOs;
@@ -14,13 +15,15 @@ public class MediaService : IMediaService
     private readonly AppDbContext _dbContext;
     private readonly IAmazonS3 _s3;
     private readonly IConfiguration _config;
+    private readonly IHttpContextAccessor _httpContextAccessor;
     private readonly string? bucket;
 
-    public MediaService(AppDbContext dbContext, IAmazonS3 s3, IConfiguration config)
+    public MediaService(AppDbContext dbContext, IAmazonS3 s3, IConfiguration config, IHttpContextAccessor httpContextAccessor)
     {
         _dbContext = dbContext;
         _s3 = s3;
         _config = config;
+        _httpContextAccessor = httpContextAccessor;
         bucket = _config["AWS:BucketName"];
     }
 
@@ -36,22 +39,35 @@ public class MediaService : IMediaService
         var request = new GetPreSignedUrlRequest
         {
             BucketName = bucket,
-            Key = media.FileName,
+            Key = mediaId,
             Verb = HttpVerb.GET,
             Expires = DateTime.UtcNow.AddMinutes(15)
         };
 
-        var url = _s3.GetPreSignedURL(request); //use async-await method
+        var url = _s3.GetPreSignedURL(request);
 
         return new() { URL = url, ContentType = media.ContentType };
     }
 
-    public async Task DeleteMediaAsync(string fileName)
+    public async Task DeleteMediaAsync(string mediaId)
     {
-        var response = await _s3.DeleteObjectAsync(new DeleteObjectRequest
+        var userId = _httpContextAccessor.HttpContext?.User.FindFirst(ClaimTypes.NameIdentifier)?.Value 
+            ?? throw new UnauthorizedAccessException("User not authenticated!");
+
+        var media = await _dbContext.Media
+            .FirstOrDefaultAsync(m => m.Id.ToString() == mediaId)
+            ?? throw new Exception("Media not found!");
+
+        if (media.UserId.ToString() != userId)
+            throw new UnauthorizedAccessException("You don't have permission to delete this media.");
+
+        _dbContext.Remove(media);
+        await _dbContext.SaveChangesAsync();
+     
+        var result = await _s3.DeleteObjectAsync(new DeleteObjectRequest
         {
             BucketName = bucket,
-            Key = fileName
+            Key = mediaId
         });
     }
 }
